+++ extensions/bland/api/schema/index.ts
@@ -1,3 +1,4 @@
 export * from './SendCall.schema'
+export * from './SendSms.schema'
 export * from './GetCallDetails.schema'
 export * from './StopActiveCall.schema'