+++ extensions/bland/CHANGELOG.md
@@ -1,1 +1,7 @@
 # CHANGELOG
+
+## Unreleased
+
+### Added
+
+- New `Send SMS` action that sends an SMS message from an agent to a user via Bland's `POST /v1/sms/send` endpoint, creating or resuming a conversation.